finished DOM functionality and multiple functions
Reworked DOM methods to display results within the playAgain fucntion
and reworked the function to display results during the round. When the
alert is confirmed the users scores are rest and the text strings are
removed from the page.

reworked the buttons.forEach method to run an alert that runs play play
Again on true. set timeout on the alert to display after DOM refresh